+++ CHANGELOG.md
@@ -8,6 +8,8 @@
   * webapp
   * sqlserver
   * sqldatabase
+* Automatically update dependsOn attribute when resources that are added depend
+  on other resources that exist in the same template
 
 ## v0.1.2